requirements.txt
flask==3.0.3 
pandas==2.0.3 
SQLAlchemy==2.0.32 
python-dotenv==1.0.1 
flask_migrate==4.0.7 
flask_dotenv==0.1.2 
pyspark==3.1.2 
databricks-cli==0.14.3 
requests==2.25.1  # To interact with Databricks REST API 